+++ lockfile_compat.go
@@ -1,0 +1,15 @@
+package storage
+
+import (
+	"github.com/containers/storage/pkg/lockfile"
+)
+
+type Locker = lockfile.Locker
+
+func GetLockfile(path string) (lockfile.Locker, error) {
+	return lockfile.GetLockfile(path)
+}
+
+func GetROLockfile(path string) (lockfile.Locker, error) {
+	return lockfile.GetROLockfile(path)
+}
