+++ pkg/lockfile/lockfile.go
@@ -1,4 +1,4 @@
-package storage
+package lockfile
 
 import (
 	"path/filepath"
