+++ pkg/lockfile/lockfile_test.go
@@ -1,4 +1,4 @@
-package storage
+package lockfile
 
 import (
 	"fmt"
